added email, sms, and email linking to TeamMemberScreen
